Optimized instruction payload size

+++ lwlog/include/details/pattern/compiler/parser.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 #include "details/pattern/pattern_instruction.h"
 #include "details/sgr_resolver.h"
@@ -18,21 +19,21 @@ namespace lwlog::details::pattern_compiler
     class parser
     {
     public:
-        void init(const char* src, std::size_t size, bool enable_color);
+        void init(const char* src, std::uint16_t size, bool enable_color);
         bool parse_sequence(pattern_bytecode::instruction_list& out, char until = '\0', bool allow_color = true);
 
     private:
-        void emit_literal_instruction(pattern_bytecode::instruction_list& out, std::uint16_t offset, std::uint16_t size);
+        void emit_literal_instruction(pattern_bytecode::instruction_list& out, std::uint16_t offset, std::uint8_t size);
         void emit_field_instruction(pattern_bytecode::instruction_list& out, pattern_bytecode::builtin_field field,
             const pattern_bytecode::alignment_info& alignment);
         void emit_custom_instruction(pattern_bytecode::instruction_list& out, 
-            std::uint16_t field_offset, std::uint16_t field_size,
+            std::uint16_t field_offset, std::uint8_t field_size,
             std::uint16_t name_offset, std::uint8_t name_size, const pattern_bytecode::alignment_info& alignment);
         void emit_sgr_begin_instruction(pattern_bytecode::instruction_list& out, std::uint8_t code);
         void emit_sgr_end_instruction(pattern_bytecode::instruction_list& out);
         void emit_sgr_begin_level_instruction(pattern_bytecode::instruction_list& out);
 
-        void flush_pending_literal(pattern_bytecode::instruction_list& out, std::size_t literal_end_offset);
+        void flush_pending_literal(pattern_bytecode::instruction_list& out, std::uint16_t literal_end_offset);
 
     private:
         bool recover_field_as_literal(pattern_bytecode::instruction_list& out);
@@ -52,10 +53,10 @@ namespace lwlog::details::pattern_compiler
 
     private:
         const char* m_src{ nullptr };
-        std::size_t m_size{ 0 };
+        std::uint16_t m_size{ 0 };
 
-        std::size_t m_pos{ 0 };
-        std::size_t m_pending_literal_begin{ 0 };
+        std::uint16_t m_pos{ 0 };
+        std::uint16_t m_pending_literal_begin{ 0 };
 
         bool m_enable_color{ false };
     };
